interface/requirements.txt
alembic==1.13.0
anyio==4.1.0
bidict==0.22.1
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
cryptography==41.0.7
deprecation==2.1.0
fastapi==0.96.1
firebase-admin==6.3.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gotrue==2.0.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.6
Jinja2==3.1.2
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msgpack==1.0.7
packaging==23.2
Pillow==10.1.0
pipdeptree==2.13.1
platformdirs==3.11.0
postgrest==0.13.0
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-multipart==0.0.5
python-socketio==5.10.0
realtime==1.0.2
redis==4.6.0
reflex==0.3.4
requests==2.31.0
rich==13.7.0
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a37
sqlmodel==0.0.8
starlette==0.27.0
starlette-admin==0.9.0
storage3==0.7.0
StrEnum==0.4.15
supabase==2.2.0
supafunc==0.3.1
tabulate==0.9.0
typer==0.9.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.20.0
watchdog==2.3.1
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.16.0
wsproto==1.2.0
